requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiosqlite>=0.19.0
pyyaml>=6.0
structlog>=24.1.0
httpx>=0.26.0
python-multipart>=0.0.6
fastapi-mcp>=0.4.0
aiodocker>=0.21.0
psutil>=5.9.0
openai>=1.40.0
anthropic>=0.34.0
google-generativeai>=0.8.0
